refactor(P28): extract the FFmpeg probe text parser into media_probe
Moves parse_probe and its private helper parse_hms out of video_decode.rs
into a new src-tauri/src/media_probe.rs. video_decode keeps one job
(spawn ffmpeg, manage decode sessions), media_probe keeps the other
(read the text ffmpeg prints). Both call sites — open_session_core and
create_optimized_media_core — are unchanged apart from the path.

Behaviour-preserving: the moved bodies are byte-identical to the originals
at 391f2e3 (verified by diff; the only change is pub(crate) visibility).
Same tuple, same error strings, same doc comments. The seven captured-output
tests moved with the parser unchanged.

The new module imports nothing from std::process and does no I/O: it takes
a &str and returns values, so its tests need no FFmpeg binary. Added seven
characterization tests for the failure paths the packet names but that had
no test — each of the five .ok_or arms plus parse_hms — pinning the EXACT
current error strings so neither this move nor a later edit can quietly
reword a diagnostic. They document the parser as it is, including where it
is lenient; they do not claim the current wording is the right wording.

Boundaries: media_probe.rs is declared in rust.desktop.media (41 -> 42
paths, 203 nonblank lines, under the 350 warn, no waiver needed).
video_decode.rs measures 2266 -> 2160; its ceiling is deliberately LEFT at
the adoption value and the reduction is recorded as nonblankLinesNow. An
extraction packet changes code, not size policy — ratcheting the allowance
down is a decision for its own packet, and doing it here would also have
given the baseline growth check 106 lines of slack (the gate's own mutation
test catches that, which is how it was found). rust.baseline.json is
untouched: that is the ratchet working.
